# Client/src/app/_generated/config/swagger-config.json
{
  "swagger": "2.0",
  "info": {
    "version": "v1",
    "title": "CLERP API v1",
    "description": "REST Web-Api-Backend for the CLERP application"
  },
  "paths": {
...
    "/api/v1/Product/{id}": {
      "get": {
        "tags": ["Product"],
        "summary": "Gets a product by the id",
        "operationId": "GetProductById",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id of the requested product",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "Product found",
            "schema": { "$ref": "#/definitions/ProductResponse" }
          },
          "404": {
            "description": "Product couldn't be found",
            "schema": { "$ref": "#/definitions/MessageResponse" }
          },
          "422": {
            "description": "Validation failed",
            "schema": { "$ref": "#/definitions/ValidationFailedResponse" }
          },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" },
          "500": {
            "description": "An unknown error occured",
            "schema": { "$ref": "#/definitions/MessageResponse" }
          }
        },
        "security": [{ "Bearer": [] }]
      },
      "delete": {
        "tags": ["Product"],
        "summary": "Deletes a product",
        "operationId": "DeleteProduct",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id of the product",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "400": {
            "description": "Product coulnd't be found",
            "schema": { "$ref": "#/definitions/BadRequestResponse" }
          },
          "200": { "description": "Product deleted successfuly" },
          "422": {
            "description": "Validation failed",
            "schema": { "$ref": "#/definitions/ValidationFailedResponse" }
          },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" },
          "500": {
            "description": "An unknown error occured",
            "schema": { "$ref": "#/definitions/MessageResponse" }
          }
        },
        "security": [{ "Bearer": [] }]
      },
      "patch": {
        "tags": ["Product"],
        "summary": "Updates a product",
        "operationId": "UpdateProduct",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "updateData",
            "in": "body",
            "description": "Data for updating a product",
            "required": false,
            "schema": { "$ref": "#/definitions/ProductUpdateRequest" }
          },
          {
            "name": "id",
            "in": "path",
            "description": "Id of the product which should be updated",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "Product successfuly updated",
            "schema": { "$ref": "#/definitions/ProductResponse" }
          },
          "400": {
            "description": "Product coulnd't be found",
            "schema": { "$ref": "#/definitions/BadRequestResponse" }
          },
          "422": {
            "description": "Validation failed",
            "schema": { "$ref": "#/definitions/ValidationFailedResponse" }
          },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" },
          "500": {
            "description": "An unknown error occured",
            "schema": { "$ref": "#/definitions/MessageResponse" }
          }
        },
        "security": [{ "Bearer": [] }]
      }
    },
    "/api/v1/ProductType": {
      "get": {
        "tags": ["ProductType"],
        "summary": "Gets all product types",
        "operationId": "GetAllProductTypes",
        "consumes": ["application/json"],
        "produces": [],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "schema": { "$ref": "#/definitions/ProductTypeGetAllResponse" }
          },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" },
          "500": {
            "description": "An unknown error occured",
            "schema": { "$ref": "#/definitions/MessageResponse" }
          }
        },
        "security": [{ "Bearer": [] }]
      },
      "post": {
        "tags": ["ProductType"],
        "summary": "Creates a new product type",
        "operationId": "CreateProductType",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "createData",
            "in": "body",
            "description": "Data for creating a new product type",
            "required": false,
            "schema": { "$ref": "#/definitions/ProductTypeCreateRequest" }
          }
        ],
        "responses": {
          "200": {
            "description": "ProductType successfuly created",
            "schema": { "$ref": "#/definitions/ProductTypeCreateResponse" }
          },
          "409": {
            "description": "Entered data conflicts with existing",
            "schema": { "$ref": "#/definitions/ConflictResponse" }
          },
          "422": {
            "description": "Validation failed",
            "schema": { "$ref": "#/definitions/ValidationFailedResponse" }
          },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" },
          "500": {
            "description": "An unknown error occured",
            "schema": { "$ref": "#/definitions/MessageResponse" }
          }
        },
        "security": [{ "Bearer": [] }]
      }
    },
    "/api/v1/ProductType/{id}": {
      "get": {
        "tags": ["ProductType"],
        "summary": "Gets a product type by the id",
        "operationId": "GetProductTypeById",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id of the product type",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "ProductType found",
            "schema": { "$ref": "#/definitions/ProductTypeResponse" }
          },
          "404": {
            "description": "ProductType couldn't be found",
            "schema": { "$ref": "#/definitions/MessageResponse" }
          },
          "422": {
            "description": "Validation failed",
            "schema": { "$ref": "#/definitions/ValidationFailedResponse" }
          },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" },
          "500": {
            "description": "An unknown error occured",
            "schema": { "$ref": "#/definitions/MessageResponse" }
          }
        },
        "security": [{ "Bearer": [] }]
      },
      "put": {
        "tags": ["ProductType"],
        "summary": "Updates a product type",
        "operationId": "UpdateProductType",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "updateData",
            "in": "body",
            "description": "Data for updating a product type",
            "required": false,
            "schema": { "$ref": "#/definitions/ProductTypeUpdateRequest" }
          },
          {
            "name": "id",
            "in": "path",
            "description": "Id of the product type",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "ProductType successfuly updated",
            "schema": { "$ref": "#/definitions/ProductTypeResponse" }
          },
          "400": {
            "description": "ProductType couldn't be found",
            "schema": { "$ref": "#/definitions/BadRequestResponse" }
          },
          "422": {
            "description": "Validation failed",
            "schema": { "$ref": "#/definitions/ValidationFailedResponse" }
          },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" },
          "500": {
            "description": "An unknown error occured",
            "schema": { "$ref": "#/definitions/MessageResponse" }
          }
        },
        "security": [{ "Bearer": [] }]
      },
      "delete": {
        "tags": ["ProductType"],
        "summary": "Deletes a product type",
        "operationId": "DeleteProductType",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id of the prodcut type",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "400": {
            "description": "ProductType coulnd't be found",
            "schema": { "$ref": "#/definitions/BadRequestResponse" }
          },
          "200": { "description": "ProductType deleted successfuly" },
          "422": {
            "description": "Validation failed",
            "schema": { "$ref": "#/definitions/ValidationFailedResponse" }
          },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" },
          "500": {
            "description": "An unknown error occured",
            "schema": { "$ref": "#/definitions/MessageResponse" }
          }
        },
        "security": [{ "Bearer": [] }]
      }
    },
    "/api/v1/ProductType/children/{id}": {
      "get": {
        "tags": ["ProductType"],
        "summary": "Gets all first level children of a product type",
        "operationId": "GetAllChildrenFromProductType",
        "consumes": ["application/json"],
        "produces": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id of the parent product type",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/ProductTypeGetAllChildrenResponse"
            }
          },
          "422": {
            "description": "Validation failed",
            "schema": { "$ref": "#/definitions/ValidationFailedResponse" }
          },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" },
          "500": {
            "description": "An unknown error occured",
            "schema": { "$ref": "#/definitions/MessageResponse" }
          }
        },
        "security": [{ "Bearer": [] }]
      }
    },
    "/api/v1/ProductType/add-parent": {
      "post": {
        "tags": ["ProductType"],
        "summary": "Adds a product type as a new parent to a product type",
        "operationId": "AddParentProductType",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "productTypeAddParenData",
            "in": "body",
            "description": "Data for adding a new parent product type",
            "required": false,
            "schema": { "$ref": "#/definitions/ProductTypeAddParentRequest" }
          }
        ],
        "responses": {
          "400": {
            "description": "Product type or parent product type coulnd't be found",
            "schema": { "$ref": "#/definitions/BadRequestResponse" }
          },
          "409": {
            "description": "Entered data conflicts with existing",
            "schema": { "$ref": "#/definitions/ConflictResponse" }
          },
          "200": { "description": "Parent successfuly added" },
          "422": {
            "description": "Validation failed",
            "schema": { "$ref": "#/definitions/ValidationFailedResponse" }
          },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" },
          "500": {
            "description": "An unknown error occured",
            "schema": { "$ref": "#/definitions/MessageResponse" }
          }
        },
        "security": [{ "Bearer": [] }]
      }
    },
    "/api/v1/ProductType/add-child": {
      "post": {
        "tags": ["ProductType"],
        "summary": "Adds a product type as a new child to a product type",
        "operationId": "AddChildProductType",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "productTypeAddChildData",
            "in": "body",
            "description": "Data for adding a new child product type",
            "required": false,
            "schema": { "$ref": "#/definitions/ProductTypeAddChildRequest" }
          }
        ],
        "responses": {
          "400": {
            "description": "Product type or child product type coulnd't be found",
            "schema": { "$ref": "#/definitions/BadRequestResponse" }
          },
          "409": {
            "description": "Entered data conflicts with existing",
            "schema": { "$ref": "#/definitions/ConflictResponse" }
          },
          "200": { "description": "Child successfuly added" },
          "422": {
            "description": "Validation failed",
            "schema": { "$ref": "#/definitions/ValidationFailedResponse" }
          },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" },
          "500": {
            "description": "An unknown error occured",
            "schema": { "$ref": "#/definitions/MessageResponse" }
          }
        },
        "security": [{ "Bearer": [] }]
      }
    },
    "/api/v1/ProductType/remove-parent": {
      "delete": {
        "tags": ["ProductType"],
        "summary": "Removes a parent product type from a product type (base)",
        "operationId": "RemoveParentProductType",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "productTypeRemoveParentData",
            "in": "body",
            "description": "Data for removing the parent from the base product type",
            "required": false,
            "schema": { "$ref": "#/definitions/ProductTypeRemoveParentRequest" }
          }
        ],
        "responses": {
          "400": {
            "description": "Parent or base product type couln't be found, or the base product type doesn't have this product type as parent",
            "schema": { "$ref": "#/definitions/BadRequestResponse" }
          },
          "200": {
            "description": "Parent successfuly removed from the product type"
          },
          "422": {
            "description": "Validation failed",
            "schema": { "$ref": "#/definitions/ValidationFailedResponse" }
          },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" },
          "500": {
            "description": "An unknown error occured",
            "schema": { "$ref": "#/definitions/MessageResponse" }
          }
        },
        "security": [{ "Bearer": [] }]
      }
    },
    "/api/v1/ProductType/remove-child": {
      "delete": {
        "tags": ["ProductType"],
        "summary": "Removes a child product type from a product type (base)",
        "operationId": "RemoveChildProductType",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "productTypeRemoveChildData",
            "in": "body",
            "description": "Data for removing the child from the base product type",
            "required": false,
            "schema": { "$ref": "#/definitions/ProductTypeRemoveChildRequest" }
          }
        ],
        "responses": {
          "400": {
            "description": "Child or base product type couln't be found, or the base product type doesn't have this product type as child",
            "schema": { "$ref": "#/definitions/BadRequestResponse" }
          },
          "200": {
            "description": "Child successfuly removed from the product type"
          },
          "422": {
            "description": "Validation failed",
            "schema": { "$ref": "#/definitions/ValidationFailedResponse" }
          },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" },
          "500": {
            "description": "An unknown error occured",
            "schema": { "$ref": "#/definitions/MessageResponse" }
          }
        },
        "security": [{ "Bearer": [] }]
      }
    },
...
  },
  "definitions": {
...
    "MessageResponse": {
      "description": "A response contianer when just a message have to be returned",
      "type": "object",
      "properties": { "message": { "type": "string" } }
    },
...
    "ConflictResponse": {
      "description": "Payload of a CLERP.API.Infrastructure.Exceptions.ConflictException",
      "type": "object",
      "properties": {
        "field": {
          "description": "Name of the property which caused the conflict",
          "type": "string",
          "readOnly": true
        },
        "message": {
          "description": "Message about the reason/background of the conflict",
          "type": "string"
        }
      }
    },
    "ValidationFailedResponse": {
      "description": "Represents the summary of all validation errors\r\nData-Format that will be returned after a failed validation",
      "type": "object",
      "properties": {
        "message": { "type": "string", "readOnly": true },
        "validationErrors": {
          "uniqueItems": false,
          "type": "array",
          "items": { "$ref": "#/definitions/ValidationError" },
          "readOnly": true
        }
      }
    },
    "ValidationError": {
      "description": "Represents a validation error, occured when validating an incoming model",
      "type": "object",
      "properties": {
        "field": { "type": "string", "readOnly": true },
        "message": { "type": "string", "readOnly": true }
      }
    },
...
    "BadRequestResponse": {
      "description": "Payload of CLERP.API.Infrastructure.Exceptions.BadRequestException",
      "type": "object",
      "properties": {
        "message": {
          "description": "Error message, which explains that something went wrong",
          "type": "string"
        }
      }
    },
...
    "ProductResponse": {
      "description": "Default response dto for a product",
      "type": "object",
      "properties": {
        "id": { "format": "uuid", "type": "string" },
        "name": { "type": "string" },
        "serialNumber": { "type": "string" },
        "state": {
          "enum": ["inStock", "reserved", "sold", "defect"],
          "type": "string"
        },
        "parentId": { "format": "uuid", "type": "string" },
        "childIds": {
          "uniqueItems": false,
          "type": "array",
          "items": { "format": "uuid", "type": "string" }
        },
        "productTypeId": { "format": "uuid", "type": "string" },
        "compartmentId": { "format": "uuid", "type": "string" },
        "billId": { "format": "uuid", "type": "string" }
      }
    },
...
    "ProductUpdateRequest": {
      "required": ["state"],
      "type": "object",
      "properties": {
        "description": { "type": "string" },
        "state": {
          "minLength": 1,
          "enum": ["inStock", "reserved", "sold", "defect"],
          "type": "string"
        },
        "compartmentId": { "format": "uuid", "type": "string" }
      }
    },
    "ProductTypeGetAllResponse": {
      "type": "object",
      "properties": {
        "productTypes": {
          "uniqueItems": false,
          "type": "array",
          "items": { "$ref": "#/definitions/ProductTypeResponse" }
        }
      }
    },
    "ProductTypeResponse": {
      "description": "Default response dto for a product type",
      "type": "object",
      "properties": {
        "id": { "format": "uuid", "type": "string" },
        "name": { "type": "string" },
        "Description": { "type": "string" },
        "ean": { "description": "European articel number", "type": "string" },
        "price": { "format": "double", "type": "number" },
        "imageBase64": {
          "description": "Image as base64 string",
          "type": "string"
        },
        "childIds": {
          "description": "All ids from the entities related as children to this product type",
          "uniqueItems": false,
          "type": "array",
          "items": { "format": "uuid", "type": "string" }
        },
        "parentIds": {
          "description": "All ids from the entities related as paretns to this product type",
          "uniqueItems": false,
          "type": "array",
          "items": { "format": "uuid", "type": "string" }
        }
      }
    },
    "ProductTypeCreateRequest": {
      "required": ["name", "ean", "price"],
      "type": "object",
      "properties": {
        "name": { "maxLength": 50, "minLength": 2, "type": "string" },
        "description": { "type": "string" },
        "ean": {
          "description": "European Article Number",
          "minLength": 1,
          "type": "string"
        },
        "price": {
          "format": "double",
          "minimum": 0.01,
          "exclusiveMinimum": true,
          "minLength": 1,
          "type": "number"
        },
        "imageBase64": {
          "description": "Image as base64 string",
          "type": "string"
        },
        "parentIds": {
          "description": "Collection with all ids of all parent product types",
          "uniqueItems": false,
          "type": "array",
          "items": { "format": "uuid", "type": "string" }
        },
        "childIds": {
          "description": "Collection with all ids of all the child product types",
          "uniqueItems": false,
          "type": "array",
          "items": { "format": "uuid", "type": "string" }
        }
      }
    },
    "ProductTypeCreateResponse": {
      "type": "object",
      "properties": {
        "productTypeId": {
          "format": "uuid",
          "description": "Id of the created product type",
          "type": "string"
        }
      }
    },
    "ProductTypeUpdateRequest": {
      "required": ["name", "ean", "price"],
      "type": "object",
      "properties": {
        "name": { "maxLength": 50, "minLength": 2, "type": "string" },
        "description": { "type": "string" },
        "ean": {
          "description": "European Article Number",
          "minLength": 1,
          "type": "string"
        },
        "price": {
          "format": "double",
          "minimum": 0.01,
          "exclusiveMinimum": true,
          "minLength": 1,
          "type": "number"
        }
      }
    },
    "ProductTypeGetAllChildrenResponse": {
      "type": "object",
      "properties": {
        "children": {
          "description": "All first level children of the product type",
          "uniqueItems": false,
          "type": "array",
          "items": { "$ref": "#/definitions/ProductTypeResponse" }
        }
      }
    },
    "ProductTypeAddParentRequest": {
      "type": "object",
      "properties": {
        "parentId": {
          "format": "uuid",
          "description": "Id of the new parent product type",
          "type": "string"
        },
        "baseId": {
          "format": "uuid",
          "description": "Id of the product type which the parent product type should be added to",
          "type": "string"
        }
      }
    },
    "ProductTypeAddChildRequest": {
      "type": "object",
      "properties": {
        "childId": {
          "format": "uuid",
          "description": "Id of the new child product type",
          "type": "string"
        },
        "baseId": {
          "format": "uuid",
          "description": "Id of the product type which the child product type should be added to",
          "type": "string"
        }
      }
    },
    "ProductTypeRemoveParentRequest": {
      "type": "object",
      "properties": {
        "parentId": {
          "format": "uuid",
          "description": "Id of the parent product type which should be removed",
          "type": "string"
        },
        "baseId": {
          "format": "uuid",
          "description": "Id of the product type which the parent should be removed from",
          "type": "string"
        }
      }
    },
    "ProductTypeRemoveChildRequest": {
      "type": "object",
      "properties": {
        "childId": {
          "format": "uuid",
          "description": "Id of the child product type which should be removed",
          "type": "string"
        },
        "baseId": {
          "format": "uuid",
          "description": "Id of the product type which the child should be removed from",
          "type": "string"
        }
      }
    },
...
  },
  "securityDefinitions": {
    "Bearer": {
      "name": "Authorization",
      "in": "header",
      "type": "apiKey",
      "description": "Please insert JWT with Bearer into field. Example: \"bearer {token}\""
    }
  },
...
}